+++ torchsparse/backend/convolution/convolution_forward_implicit_gemm_cuda.h
@@ -1,10 +1,11 @@
 #include <ATen/Operators.h>
+#include <cstdint>
 #include <torch/all.h>
 #include <torch/library.h>
 
 at::Tensor conv_forward_implicit_gemm_cuda(const at::Tensor &_in_feats,
                                            const at::Tensor &_kernel,
                                            const at::Tensor &_out_in_map,
-                                           int num_out_feats,
-                                           int num_out_channels,
+                                           int64_t num_out_feats,
+                                           int64_t num_out_channels,
                                            bool allow_tf32, bool allow_fp16);
