+++ torchsparse/backend/convolution/convolution_forward_implicit_gemm_sorted_cuda.h
@@ -1,4 +1,5 @@
 #include <ATen/Operators.h>
+#include <cstdint>
 #include <torch/all.h>
 #include <torch/library.h>
 
@@ -7,5 +8,5 @@ at::Tensor conv_forward_implicit_gemm_sorted_cuda(
                        const at::Tensor& _in_feats, const at::Tensor& _kernel,
                        const at::Tensor& _out_in_map, const at::Tensor& _reduced_mask,
                        const at::Tensor& _reorder_loc,
-                       int num_out_feats, int num_out_channels,
+                       int64_t num_out_feats, int64_t num_out_channels,
                        bool allow_tf32, bool allow_fp16);
